fix(zfb): make zfb/frontmatter subpath fs-free [H7]
The frontmatter subpath was a re-export shim that pointed at content.ts:

    export { parseFrontmatter } from "./content.js";

content.ts has a top-level import of node:fs/promises (for the
getCollection filesystem fallback), so importing @takazudo/zfb/frontmatter
in a Workers / edge bundle still pulled node:fs/promises into the bundle
graph — defeating the BCI-3 promise that the subpath stays fs-free.

Move parseFrontmatter (and ParsedFrontmatter) into frontmatter.ts as
the canonical home. content.ts now imports from ./frontmatter.js and
re-exports the names so existing zfb/content consumers keep working.

Added a static smoke test that walks the import graph reachable from
frontmatter.ts and asserts no module on the path imports node:fs* —
the test fails fast if a future refactor accidentally re-introduces
the dependency.

Refs PR #64 review (issue #80).
